changes #2

+++ src/main/java/com/beganov/studentservice/controller/StudentController.java
@@ -1,6 +1,7 @@
 package com.beganov.studentservice.controller;
 
-import com.beganov.studentservice.dto.StudentDto;
+import com.beganov.studentservice.dto.StudentRequest;
+import com.beganov.studentservice.dto.StudentResponse;
 import com.beganov.studentservice.service.StudentService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -20,34 +21,43 @@ public class StudentController {
     }
 
     @GetMapping
-    public ResponseEntity<List<StudentDto>> getAllStudents() {
-        List<StudentDto> list = studentService.getAllStudents();
-        return ResponseEntity.ok(list);
+    public ResponseEntity<List<StudentResponse>> getAllStudents() {
+        List<StudentResponse> list = studentService.getAllStudents();
+        return list.isEmpty() ? ResponseEntity.status(HttpStatus.NOT_FOUND).build() : ResponseEntity.ok(list);
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<StudentDto> getStudentById(@PathVariable Long id) {
-        Optional<StudentDto> studentDto = studentService.getStudentById(id);
-        return studentDto.map(ResponseEntity::ok)
-                .orElseGet(() -> ResponseEntity.status(HttpStatus.NOT_FOUND).build());
+    public ResponseEntity<?> getStudentById(@PathVariable Long id) {
+        Optional<StudentResponse> response = studentService.getStudentById(id);
+        return response
+                .<ResponseEntity<?>>map(ResponseEntity::ok)
+                .orElseGet(() -> ResponseEntity.status(HttpStatus.NOT_FOUND)
+                        .body("Студент с ID " + id + " не найден"));
     }
 
     @PutMapping
-    public ResponseEntity<StudentDto> saveStudent(@RequestBody StudentDto studentDto) {
-        StudentDto createdStudent = studentService.saveStudent(studentDto);
+    public ResponseEntity<StudentResponse> saveStudent(@RequestBody StudentRequest request) {
+        StudentResponse createdStudent = studentService.saveStudent(request);
         return ResponseEntity.status(HttpStatus.CREATED).body(createdStudent);
     }
 
     @PostMapping("/{id}")
-    public ResponseEntity<StudentDto> updateStudent(@PathVariable Long id, @RequestBody StudentDto studentDto) {
-        Optional<StudentDto> updatedStudent = studentService.updateStudent(id, studentDto);
-       return updatedStudent.map(ResponseEntity::ok)
-               .orElseGet(() -> ResponseEntity.status(HttpStatus.NOT_FOUND).build());
+    public ResponseEntity<?> updateStudent(@PathVariable Long id, @RequestBody StudentRequest request) {
+        Optional<StudentResponse> updatedStudent = studentService.updateStudent(id, request);
+        if (updatedStudent.isPresent()) {
+            return ResponseEntity.ok(updatedStudent.get());
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Студент с ID " + id + " не существует");
+        }
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteStudent(@PathVariable Long id) {
-        studentService.deleteStudent(id);
-        return ResponseEntity.noContent().build();
+    public ResponseEntity<String> deleteStudent(@PathVariable Long id) {
+        if (studentService.getStudentById(id).isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Студент с ID " + id + " не существует");
+        } else {
+            studentService.deleteStudent(id);
+            return ResponseEntity.status(HttpStatus.OK).body("Студен с ID " + id + " удален" );
+        }
     }
 }
